+++ tests/core/recipes/orca_recipes/test_orca_recipes.py
@@ -213,3 +213,25 @@ def test_fairchem_omol(tmp_path, monkeypatch):
         output["parameters"]["orcasimpleinput"]
         == "allpop def2-tzvpd def2/j defgrid3 diis engrad nonbo nonpa normalconv nososcf nousesym rijcosx wb97m-v xyzfile"
     )
+    assert "%pal nprocs 12 end" in output["parameters"]["orcablocks"]
+    assert (
+        Path(output["dir_name"], "def2-tzvpd.bas").exists()
+        or Path(output["dir_name"], "def2-tzvpd.bas.gz").exists()
+    )
+
+
+@pytest.mark.skipif(not has_fairchem_omol, reason="fairchem not installed")
+def test_fairchem_omol_custom_inputs(tmp_path, monkeypatch):
+    from quacc.recipes.orca.fairchem import omol_static_job
+
+    monkeypatch.chdir(tmp_path)
+
+    orcasimpleinput = ["TightSCF"]
+    output = omol_static_job(
+        molecule("H2"), orcasimpleinput=orcasimpleinput, orcablocks=[], nprocs=2
+    )
+    assert orcasimpleinput == ["TightSCF"]
+    assert "rijcosx" not in output["parameters"]["orcasimpleinput"]
+    assert "tightscf" in output["parameters"]["orcasimpleinput"]
+    assert "%pal nprocs 2 end" in output["parameters"]["orcablocks"]
+    assert "GTOName" not in output["parameters"]["orcablocks"]